requirements.txt
addict~=2.4
celery~=5.4
connexion[swagger-ui,flask,uvicorn]~=3.0
#cryptography~=42.0
flask-authz~=2.6
Flask-PyMongo~=2.3
pydantic~=2.7
#PyJWT~=2.4
pymongo~=4.7
#PyYAML~=6.0
#requests~=2.31
#toml~=0.10
#typing~=3.7
